+++ CHANGELOG.md
@@ -1,3 +1,26 @@
+# Next
+
+* Add warning if the lost items section of the postmaster has 20 items.
+* Stat bars are more accurately sized.
+* Add vendor progress
+* Add prestige level with xp bar under characters to replace normal xp bar after level 40.
+* Add warning if the lost items section of the postmaster has 20 items.
+* It is no longer possible to choose column sizes that cause the vault to disappear.
+* The Vault now has a character-style header, and can have loadouts applied to it. Full-ness of each vault is displayed below the vault header.
+* New option to restore all the items that were in your inventory before applying a loadout, rather than just the equipped ones.
+* You can now undo multiple loadouts, going backwards in time.
+
+
+# 3.6.1
+
+* Removed the "Only blues" option in the infusion fuel finder, because it wasn't necessary.
+* Engram searches and the engram loadout features won't mistake Candy Engrams for real engrams.
+* Items in the Postmaster include their type in the move popup, so they're easier to distinguish.
+* Sometimes equipping loadouts would fail to equip one of your exotics. No more!
+* Add an 'is:infusable' search filter.
+* Add 'is:intellect', 'is:discipline', 'is:strength' search filters for armor.
+* XP Progress on bar items
+
 # 3.6.0
 
 * Bring back the infusion dialog as an Infusion Fuel Finder. It doesn't do as much as it used to, but now it's optimized for quickly finding eligable infusion items.